Rivista gestione emissione token e risposta modificata per mimare quella oauth2

+++ src/main/java/it/loreluc/sagraservice/config/SagraSettings.java
@@ -22,5 +22,5 @@ public class SagraSettings {
      * Duranta in minuti del token
      */
     @NotNull @Min(0)
-    private Long tokenExpire;
+    private Integer tokenExpire;
 }
